Unit tests for hyperjump functions
Hyperjump is supposed to support both bash and zsh, but it's easy for
people who use one to break support for the other. By adding some tests,
we can more easily ensure that support for both shells is maintained.

Assertions for autocompletion are skipped in zsh because they don't run
properly. I guess the bashcompinit is doing some magic there. The tests
still run to catch syntax errors, as it seems to work as long as we
don't have any of those.

This also adds a .travis.yml so the tests can be run automatically on
Travis for both shells and both osx and linux.

You can run the tests by installing shunit2 and running bash tests.sh
or zsh -y tests.sh.

Note that this allows the db location to be manipulated via setting the
HYPERJUMP_DB shell parameter. This is necessary to run tests locally
without overwriting the user's db.